After refactoring. Removed inMemory

+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exceptions.NotFoundEntityByIdException;
 import ru.yandex.practicum.filmorate.model.Director;
@@ -7,6 +8,7 @@ import ru.yandex.practicum.filmorate.storage.director.DirectorStorage;
 
 import java.util.List;
 
+@Slf4j
 @Service
 public class DirectorService {
     private final DirectorStorage directorDbStorage;
@@ -20,7 +22,11 @@ public class DirectorService {
     }
 
     public Director getDirectorById(long id) {
-        return directorDbStorage.getDirectorById(id);
+        return directorDbStorage.getDirectorById(id).orElseThrow(() -> {
+                    log.warn("DirectorService, getDirectorById, режисер с id {} не найден", id);
+                    throw new NotFoundEntityByIdException("Режиссер не найден", id);
+                }
+        );
     }
 
     public Director postDirector(Director director) {
@@ -29,12 +35,17 @@ public class DirectorService {
 
     public Director putDirector(Director director) {
         if (!directorDbStorage.containsDirectorById(director.getId())) {
+            log.warn("DirectorService, putDirector, режисер с id {} не найден", director.getId());
             throw new NotFoundEntityByIdException("Режиссер не найден", director.getId());
         }
         return directorDbStorage.putDirector(director);
     }
 
     public void deleteDirector(long id) {
+        if (!directorDbStorage.containsDirectorById(id)) {
+            log.warn("DirectorService, deleteDirector, режисер с id {} не найден", id);
+            throw new NotFoundEntityByIdException("Режиссер не найден по id", id);
+        }
         directorDbStorage.deleteDirector(id);
     }
